Create a separate template_utils library.
